[CDG-66][CDG-67] - Refactor dashboard page in order to use hooks
